Fix KSkillData serializer skill id width to legacy int

+++ KncWX2Server.Net/src/KncWX2Server.Core/Common/Serialization/KSkillDataSerialization.cs
@@ -13,7 +13,7 @@ public static class KSkillDataSerialization
     {
         value = default;
 
-        if (!serializer.Get(out short skillId)
+        if (!serializer.Get(out int skillId)
             || !serializer.Get(out byte skillLevel))
             return false;
